put usePricing in a module
